Версия 3.0.9 - экспорт/ипорт данных в формате SQL

+++ api/backup_holy.php
@@ -3,11 +3,12 @@
 /**
  * Производит бэкап текущей базы данных в файл. По-умолчанию конвертирует данные из кодировки win-1251 в utf-8
  * 
- * @param string $tables  <p>список таблиц. если не указан, сохраняются ВСЕ.</p>
+ * @param string $tables  <p>список таблиц (массивом или через запятую). Если не указан, сохраняются ВСЕ.</p>
  * @param string $skip_id  <p>не сохранять id элементов (не обязательное)</p>
  * @param string $file_out_name  <p>имя файла на выходе, учитывать DOCUMENT_ROOT (не обязательное)</p>
+ * @param string $skip_create_tables  <p>не сохранять команды создания таблиц (не обязательное)</p>
  */
-function backup_database_tables($tables = "*", $skip_id = false, $file_out_name = "") {
+function backup_database_tables($tables = "*", $skip_id = false, $file_out_name = "", $skip_create_tables = false) {
     $return = "";
     $razdel = "\n";
     $razdel2 = $razdel;
@@ -26,13 +27,20 @@ function backup_database_tables($tables = "*", $skip_id = false, $file_out_name 
     foreach ($tables as $table) {
         $result = mysql_query('SELECT * FROM ' . $table);
         $num_fields = mysql_num_fields($result);
-        $return.= 'DROP TABLE IF EXISTS ' . $table . ';';
-        $row2 = mysql_fetch_row(mysql_query('SHOW CREATE TABLE ' . $table));
-        $row2 = preg_replace("/\\n/", "", $row2);
-        $return.= $razdel2 . "" . $row2[1] . ";";
+        if (!$skip_create_tables) {
+            $return.= 'DROP TABLE IF EXISTS ' . $table . ';';
+            $row2 = mysql_fetch_row(mysql_query('SHOW CREATE TABLE ' . $table));
+            $row2 = preg_replace("/\\n/", "", $row2);
+            $return.= $razdel2 . "" . $row2[1] . ";";
+        };
+        $line1 = 0;
         for ($i = 0; $i < $num_fields; $i++) {
             while ($row = mysql_fetch_row($result)) {
-                $return.= $razdel . 'INSERT INTO ' . $table . ' VALUES(';
+                $line1++;
+                if (($line1 > 1) || (!$skip_create_tables)) {
+                    $return.= $razdel;
+                }
+                $return.= 'INSERT INTO ' . $table . ' VALUES(';
                 for ($j = 0; $j < $num_fields; $j++) {
                     $row[$j] = mysql_escape_string($row[$j]);
                     if (($skip_id) && ($j == 0))
@@ -49,7 +57,8 @@ function backup_database_tables($tables = "*", $skip_id = false, $file_out_name 
                 $return.= ");";
             }
         }
-        $return.=$razdel3;
+        if (!$skip_create_tables)
+            $return.=$razdel3;
     }
     //save the file
     global $_CONFIG;
@@ -95,9 +104,7 @@ function GetTableDatas($array_of_table = array()) {
                 $tables_info[$table] = $row;
             };
     return $tables_info;
-}
-
-;
+};
 ?>
 <?
 
@@ -108,8 +115,6 @@ function GetTableDatas($array_of_table = array()) {
  * @param array $conv  <p>Переконвертировать файл перед импортом. [0] - ИЗ какой кодировки, [1] - В какую кодировку</p>
  */
 function ImportSQL($path, $conv = array()) {
-    if (strpos($path, $_SERVER['DOCUMENT_ROOT']) === FALSE)
-        $path = $_SERVER['DOCUMENT_ROOT'] . $path;
 
     $f_data = file($path);
